Simplify logic to handle client / server binaries
+ update versions json files
