+++ versions-client.json
@@ -1,0 +1,74 @@
+{
+  "2021-04-22T17-40-00Z": {
+    "darwin-amd64": "741f8839be58740535d15f1e96532af94e5d2b44f1889dd5b6ab6071b6b24afb",
+    "darwin-arm64": "aa9ad1336555bf7cd80adf4a7660aa8e7fe54c6ee9749f5b26f719ad20ce4411",
+    "linux-amd64": "88f4e589d86dbf3031f9319fc32fd3113f3853ff91e3a5e2cce8c16d39601a67",
+    "linux-arm": "5f62387aaced0212d9127d94cde1c5147fd86f05bed39d6c86f131ab2d745bea",
+    "linux-arm64": "23c98c585da3b3c6587445ce85e949d2618f80ab187641059007757ce3167df3",
+    "windows-amd64": "01a849d4387409ec6f257e19aef3c1c9dbd58fa655326d0a1f36276f883ad677"
+  },
+  "2021-05-12T01-29-06Z": {
+    "darwin-amd64": "f4950a8980d333170d4ba3b1551446bb42f4728c4ed98952c10069865c8c83ba",
+    "darwin-arm64": "65b488ae51c6b5279de1dd0ab7834bf42a658e62168cd3854433b4177eb95b96",
+    "linux-amd64": "c6156d79af7d8b3b65a3a9ec89a027e3f617e212d73a495cebb18e038459c1aa",
+    "linux-arm": "581d98b1dc681661ba5f5810492cb6dc3f9a49b0d2e58efe0f5452817bef5458",
+    "linux-arm64": "16e2f005f3804f17d2cb55641b58eaf79c98abb25759b8ff1db356c1cc009a3d",
+    "windows-amd64": "87ffe64e3b8be56a4aa897dc400978f3fdc2e8de03a36841e614d417cc2fcda6"
+  },
+  "2021-05-12T03-10-11Z": {
+    "darwin-amd64": "b3829df3b8a47407d1ee792686f244943893d838271cb4bbe7fdb2ae6704a4ca",
+    "darwin-arm64": "2a8272deea7454e96a50a4b2b0e31017ed2c930086b09afd44141fb0c83700c3",
+    "linux-amd64": "d1c322cfd029cca13a44e09e6b9eaa9146186bfd325a5e4d2cfd40aed5d42572",
+    "linux-arm": "7b31b5cb3eee78b72f49f02dfbe40a7c0ba0818ca4363de179627be76d68d212",
+    "linux-arm64": "d925aff403729785fdd0066d70a03ea35ab7a0c775c1730b7ca8ac579b9ff05f",
+    "windows-amd64": "a7e99f920bcfabc84a34da8a0f603b5e50bc87923ae65ef1a8de8e70707e8b97"
+  },
+  "2021-05-18T03-39-44Z": {
+    "darwin-amd64": "d7ff1a7c49941fbbf5b8f4fb7da2f7fde1e68b3b061acdf8a2dd5a3934b95e0e",
+    "darwin-arm64": "efebf194d9a08cb1afd1a40432293ff748d6e1fedbc645ee3d1dbe4fa3ccb1cf",
+    "linux-amd64": "e1612024c538c7289cc007b688397c7ed3dcdc37d180e15f90703571c0d59836",
+    "linux-arm": "f5310ca295c06219fc43e12240f761b4dc88473048bc389e72f9d606602407b0",
+    "linux-arm64": "48f2e7a5595eb3d900ecaa69acea35174cb55e7c29dcd5c8c7a82db7d4e59ef5",
+    "windows-amd64": "5b330b036d58b401a8b96e4a2ca062ca26083d5eb5f72a13c9209cf2bb84c1ad"
+  },
+  "2021-05-26T19-19-26Z": {
+    "darwin-amd64": "5bb933786b60bbdeea8535f3642c69b048a24516c53cd29c4167e273e78ffc20",
+    "darwin-arm64": "bdc9a6cc9c271caf7eecdf40af15c591ff1e08a24b2c8b7f95453c83273d0905",
+    "linux-amd64": "f2fcdc16f1cd806c554a309f19ecefee1684f16fdb6d903eea6b4a52f7c5c724",
+    "linux-arm": "edcce80b515770871026c1d5f15eff6cebc7060f88261813f39cdb8d7c5d877d",
+    "linux-arm64": "2bfb06000bf7aabed6e0cf0d2d24731c1b73de5ca437b2920f991d92553b1423",
+    "windows-amd64": "109258a37aae63d29939fa39d2c3a4f681faacf2b66e023241d761aad4627ad4"
+  },
+  "2021-06-08T01-29-37Z": {
+    "darwin-amd64": "6fd86067eeca2ccff0c06c10519af230ec305710a35f5771b48a4994ba4391a1",
+    "darwin-arm64": "c66ae95323b435dee16dc3d0a125bd9f38ac2db2974d5ebe0bda8a639f439ff7",
+    "linux-amd64": "1870e6b0565fffe2579278755925ece884766fadea8e876b5f041a9eb33d6d4c",
+    "linux-arm": "b6f21603dd2adcdd6bc8994aa092e6df79b12bc3e6ef49c5f4bbc75276bf350a",
+    "linux-arm64": "1d53d0a8625be03d117490fb0967abf21651c4e23c83fea60cb88ce70da7a195",
+    "windows-amd64": "8e591113f10e56117838b6d77d618eac986565c3c64b596e3157a6e81e6af2dc"
+  },
+  "2021-06-13T17-48-22Z": {
+    "darwin-amd64": "52851e4ed44cdda23b216700a5839b64e6e6748c8e75369bee86896da1a1aa27",
+    "darwin-arm64": "f1281884f8ec9a52e101f1c0df8b3bf3fafd4265753e6cbb13aa78a8c164ef7f",
+    "linux-amd64": "83be163227d125e260025e9768924d3fa85d9c299aa58b40c6701b9444789e2d",
+    "linux-arm": "9e77d7169a50cf4c3423f83aee7a5321c7ff6cc89ec5117179192470c26a040a",
+    "linux-arm64": "9f432c5e706e4a373ecf0634b2af8c8c7799bcf5da72bdd92e5440a643477591",
+    "windows-amd64": "fbff329016161160f8ade3f5232233f3e3d7d3a78770e124d04c9603118777ea"
+  },
+  "2021-07-27T06-46-19Z": {
+    "darwin-amd64": "eb697f202f073901f8936f001bffcce709dc431fd0c70793cdbb683b895b63ea",
+    "darwin-arm64": "594780e8733699353efa7715c624009c27c116aeb0f463141a01b7c284bfad12",
+    "linux-amd64": "0df81285771e12e16a0c4c2f5e0ebc700e66abb8179013cc740d48b0abad49be",
+    "linux-arm": "253ba66ab3806666913233fad5c56e8ac7c445a36d96d449c0fa34f35f0fbd93",
+    "linux-arm64": "c961239014104e3948f80c6602423a74aa26f685f2df32199ce0c9d94e440926",
+    "windows-amd64": "5a377f80b9f5770e98e7e9c2a803ef77bae00cac4e4cd3cca94f5be74c1a39e1"
+  },
+  "2021-09-02T09-21-27Z": {
+    "darwin-amd64": "7cde9b771e4e50db6becce8d1c353cd256319a527f61eb08d616834e84be3034",
+    "darwin-arm64": "b419926c794d83002272bc76d1cdca82b824a7533711551d8ad4c13f5b6e305c",
+    "linux-amd64": "43f5a8ea5b0387c56f6c81f42bf6cc05942a5956eba9c088d0fe351b3e990b07",
+    "linux-arm": "713be3eddf9b41a8a245d82e681ade8e6c406764420f5ce12ad62649f4aa3802",
+    "linux-arm64": "b12f02597c8e26e5017bee462b834ff74e06f77c94bc8a3ab5dda0f0fa0c21f8",
+    "windows-amd64": "f87e2882747e0e456c9696c244822d9ea0a56c0b53fc4766eb15a38c57e5bd3f"
+  }
+}
